requirements.txt
# Super Agent to Skill - 依赖清单
# Python 3.10+

# ==========================================
# 核心依赖
# ==========================================
httpx>=0.25.0
pyyaml>=6.0
pydantic>=2.0.0
jsonschema>=4.0.0
python-dotenv>=1.0.0

# ==========================================
# LLM 客户端 (OpenAI 兼容 API)
# ==========================================
openai>=1.0.0

# ==========================================
# Web 框架
# ==========================================
fastapi>=0.109.0
uvicorn>=0.27.0

# ==========================================
# MCP 相关 (可选)
# ==========================================
# mcp>=0.9.0

# ==========================================
# 开发依赖 (可选)
# ==========================================
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
